refactor core modules, add tests and lint
