+++ .gitignore
@@ -1,2 +1,2 @@
 dist/
-my-html-page/
+node_modules/
