+++ src/state.js
@@ -4,15 +4,8 @@ export const state = {
   fogEnabled: true,
   players: [],
   neutral: { units: [] },
-  isBlocked: (...args) => false,
+  isBlocked: () => false,
   rand: (min = 0, max = 1) => Math.random() * (max - min) + min,
 };
 
 export default state;
-
-['paused', 'muted', 'fogEnabled', 'players', 'neutral', 'isBlocked', 'rand'].forEach(key => {
-  Object.defineProperty(globalThis, key, {
-    get: () => state[key],
-    set: v => { state[key] = v; },
-  });
-});
